requirements.txt
bcrypt
kivy
pymongo
python-dotenv
kivymd
